CinemaTicketSystem
This project will improve the functionality of a cinema. The staff members will have two options prompted to them when they run the program, one is to update the data about the movies, dates, times and price, and the other is to make reservations for guests, which will include things like what day they want to reserve the seats for, which movie, at what venue. Each day of the week is going to have options for different movies at different times, at different venues (if the cinema has more than one).

This will get rid of the problem with people sitting in seats that weren’t reserved by them. It will improve the speed and quality of the service the cinema would offer without it, such as removing unnecessary waiting in lines, providing the ability to reserve seats online or over the phone, everything can get recorded easily and monitored with the program. The program will also provide a wide variety of options for snacks and/or drinks that will be billed to the same receipt as the movie ticket.

We will be using the Java programming language to design this project. We plan on building the whole menu system using the JavaFX software platform, it's going to consist of fxml files and their controllers, we also plan on making a database with MySQL and connecting it with the Java code for full program functionality.

The project is designed with the goal to make everyone’s life easier, for the staff members as well as the customers.